add verifyHtmlResponse test case
