+++ seatunnel-engine/seatunnel-engine-server/src/main/java/org/apache/seatunnel/engine/server/rest/RestHttpGetCommandProcessor.java
@@ -17,60 +17,36 @@
 
 package org.apache.seatunnel.engine.server.rest;
 
-import org.apache.seatunnel.shade.com.fasterxml.jackson.databind.node.ArrayNode;
-
 import org.apache.seatunnel.common.exception.SeaTunnelRuntimeException;
 import org.apache.seatunnel.common.utils.FileUtils;
 import org.apache.seatunnel.common.utils.JsonUtils;
 import org.apache.seatunnel.engine.server.NodeExtension;
 import org.apache.seatunnel.engine.server.log.FormatType;
 import org.apache.seatunnel.engine.server.log.Log4j2HttpGetCommandProcessor;
-import org.apache.seatunnel.engine.server.operation.GetClusterHealthMetricsOperation;
 import org.apache.seatunnel.engine.server.rest.service.JobInfoService;
+import org.apache.seatunnel.engine.server.rest.service.LogService;
 import org.apache.seatunnel.engine.server.rest.service.OverviewService;
 import org.apache.seatunnel.engine.server.rest.service.RunningThreadService;
 import org.apache.seatunnel.engine.server.rest.service.SystemMonitoringService;
 import org.apache.seatunnel.engine.server.rest.service.ThreadDumpService;
-import org.apache.seatunnel.engine.server.utils.NodeEngineUtil;
 
 import org.apache.commons.lang3.StringUtils;
-import org.apache.logging.log4j.LogManager;
-import org.apache.logging.log4j.core.LoggerContext;
-import org.apache.logging.log4j.core.config.builder.api.Component;
-import org.apache.logging.log4j.core.config.builder.impl.BuiltConfiguration;
-import org.apache.logging.log4j.core.config.properties.PropertiesConfiguration;
-import org.apache.logging.log4j.core.lookup.StrSubstitutor;
 
-import com.hazelcast.cluster.Address;
-import com.hazelcast.cluster.Cluster;
-import com.hazelcast.cluster.Member;
 import com.hazelcast.internal.ascii.TextCommandService;
 import com.hazelcast.internal.ascii.rest.HttpCommandProcessor;
 import com.hazelcast.internal.ascii.rest.HttpGetCommand;
 import com.hazelcast.internal.ascii.rest.RestValue;
-import com.hazelcast.internal.json.JsonArray;
-import com.hazelcast.internal.json.JsonObject;
 import com.hazelcast.internal.util.JsonUtil;
 import com.hazelcast.internal.util.StringUtil;
-import com.hazelcast.jet.datamodel.Tuple3;
 import com.hazelcast.spi.impl.NodeEngineImpl;
 import io.prometheus.client.exporter.common.TextFormat;
 
-import java.io.ByteArrayOutputStream;
 import java.io.File;
 import java.io.IOException;
-import java.io.InputStream;
 import java.io.StringWriter;
-import java.lang.reflect.Field;
-import java.net.HttpURLConnection;
-import java.net.URL;
 import java.nio.charset.StandardCharsets;
-import java.util.ArrayList;
 import java.util.Arrays;
-import java.util.List;
 import java.util.Map;
-import java.util.Set;
-import java.util.concurrent.ExecutionException;
 import java.util.stream.Collectors;
 
 import static com.hazelcast.internal.ascii.rest.HttpStatusCode.SC_400;
@@ -99,6 +75,7 @@ public class RestHttpGetCommandProcessor extends HttpCommandProcessor<HttpGetCom
     private SystemMonitoringService systemMonitoringService;
     private ThreadDumpService threadDumpService;
     private RunningThreadService runningThreadService;
+    private LogService logService;
 
     public RestHttpGetCommandProcessor(TextCommandService textCommandService) {
 
@@ -109,6 +86,7 @@ public class RestHttpGetCommandProcessor extends HttpCommandProcessor<HttpGetCom
         this.systemMonitoringService = new SystemMonitoringService(nodeEngine);
         this.threadDumpService = new ThreadDumpService(nodeEngine);
         this.runningThreadService = new RunningThreadService(nodeEngine);
+        this.logService = new LogService(nodeEngine);
     }
 
     public RestHttpGetCommandProcessor(
@@ -120,6 +98,11 @@ public class RestHttpGetCommandProcessor extends HttpCommandProcessor<HttpGetCom
         this.original = log4j2HttpGetCommandProcessor;
         this.nodeEngine = this.textCommandService.getNode().getNodeEngine();
         this.overviewService = new OverviewService(nodeEngine);
+        this.jobInfoService = new JobInfoService(nodeEngine);
+        this.systemMonitoringService = new SystemMonitoringService(nodeEngine);
+        this.threadDumpService = new ThreadDumpService(nodeEngine);
+        this.runningThreadService = new RunningThreadService(nodeEngine);
+        this.logService = new LogService(nodeEngine);
     }
 
     @Override
@@ -204,41 +187,6 @@ public class RestHttpGetCommandProcessor extends HttpCommandProcessor<HttpGetCom
                 command, systemMonitoringService.getSystemMonitoringInformationJsonValues());
     }
 
-    private JsonArray getSystemMonitoringInformationJsonValues() {
-        Cluster cluster = textCommandService.getNode().hazelcastInstance.getCluster();
-
-        Set<Member> members = cluster.getMembers();
-        JsonArray jsonValues =
-                members.stream()
-                        .map(
-                                member -> {
-                                    Address address = member.getAddress();
-                                    String input = null;
-                                    try {
-                                        input =
-                                                (String)
-                                                        NodeEngineUtil.sendOperationToMemberNode(
-                                                                        nodeEngine,
-                                                                        new GetClusterHealthMetricsOperation(),
-                                                                        address)
-                                                                .get();
-                                    } catch (InterruptedException | ExecutionException e) {
-                                        logger.severe("get system monitoring information fail", e);
-                                    }
-                                    String[] parts = input.split(", ");
-                                    JsonObject jobInfo = new JsonObject();
-                                    Arrays.stream(parts)
-                                            .forEach(
-                                                    part -> {
-                                                        String[] keyValue = part.split("=");
-                                                        jobInfo.add(keyValue[0], keyValue[1]);
-                                                    });
-                                    return jobInfo;
-                                })
-                        .collect(JsonArray::new, JsonArray::add, JsonArray::add);
-        return jsonValues;
-    }
-
     private void handleRunningJobsInfo(HttpGetCommand command) {
         this.prepareResponse(command, jobInfoService.getRunningJobsJson());
     }
@@ -291,13 +239,7 @@ public class RestHttpGetCommandProcessor extends HttpCommandProcessor<HttpGetCom
         }
     }
 
-    private PropertiesConfiguration getLogConfiguration() {
-        LoggerContext context = (LoggerContext) LogManager.getContext(false);
-        return (PropertiesConfiguration) context.getConfiguration();
-    }
-
-    private void getAllNodeLog(HttpGetCommand httpGetCommand, String uri)
-            throws NoSuchFieldException, IllegalAccessException {
+    private void getAllNodeLog(HttpGetCommand httpGetCommand, String uri) {
 
         // Analysis uri, get logName and jobId param
         String param = getParam(uri);
@@ -305,73 +247,24 @@ public class RestHttpGetCommandProcessor extends HttpCommandProcessor<HttpGetCom
         String logName = isLogFile ? param : StringUtils.EMPTY;
         String jobId = !isLogFile ? param : StringUtils.EMPTY;
 
-        String logPath = getLogPath();
+        String logPath = logService.getLogPath();
         if (StringUtils.isBlank(logPath)) {
             logger.warning(
                     "Log file path is empty, no log file path configured in the current configuration file");
             httpGetCommand.send404();
             return;
         }
-        JsonArray systemMonitoringInformationJsonValues =
-                getSystemMonitoringInformationJsonValues();
 
         if (StringUtils.isBlank(logName)) {
-            StringBuffer logLink = new StringBuffer();
-            ArrayList<Tuple3<String, String, String>> allLogNameList = new ArrayList<>();
-
-            systemMonitoringInformationJsonValues.forEach(
-                    systemMonitoringInformation -> {
-                        String host = systemMonitoringInformation.asObject().get("host").asString();
-                        int port =
-                                Integer.valueOf(
-                                        systemMonitoringInformation
-                                                .asObject()
-                                                .get("port")
-                                                .asString());
-                        String url = "http://" + host + ":" + port + CONTEXT_PATH;
-                        String allName = sendGet(url + GET_ALL_LOG_NAME);
-                        logger.fine(String.format("Request: %s , Result: %s", url, allName));
-                        ArrayNode jsonNodes = JsonUtils.parseArray(allName);
-
-                        jsonNodes.forEach(
-                                jsonNode -> {
-                                    String fileName = jsonNode.asText();
-                                    if (StringUtils.isNotBlank(jobId)
-                                            && !fileName.contains(jobId)) {
-                                        return;
-                                    }
-                                    allLogNameList.add(
-                                            Tuple3.tuple3(
-                                                    host + ":" + port,
-                                                    url + GET_LOGS + "/" + fileName,
-                                                    fileName));
-                                });
-                    });
             FormatType formatType = getFormatType(uri);
             switch (formatType) {
                 case JSON:
-                    JsonArray jsonArray =
-                            allLogNameList.stream()
-                                    .map(
-                                            tuple -> {
-                                                JsonObject jsonObject = new JsonObject();
-                                                jsonObject.add("node", tuple.f0());
-                                                jsonObject.add("logLink", tuple.f1());
-                                                jsonObject.add("logName", tuple.f2());
-                                                return jsonObject;
-                                            })
-                                    .collect(JsonArray::new, JsonArray::add, JsonArray::add);
-                    this.prepareResponse(httpGetCommand, jsonArray);
+                    this.prepareResponse(httpGetCommand, logService.allNodeLogFormatJson(jobId));
                     return;
                 case HTML:
                 default:
-                    allLogNameList.forEach(
-                            tuple ->
-                                    logLink.append(
-                                            buildLogLink(
-                                                    tuple.f1(), tuple.f0() + "-" + tuple.f2())));
-                    String logContent = buildWebSiteContent(logLink);
-                    this.prepareResponse(httpGetCommand, getRestValue(logContent));
+                    this.prepareResponse(
+                            httpGetCommand, getRestValue(logService.allNodeLogFormatHtml(jobId)));
             }
         } else {
             prepareLogResponse(httpGetCommand, logPath, logName);
@@ -409,69 +302,14 @@ public class RestHttpGetCommandProcessor extends HttpCommandProcessor<HttpGetCom
         return restValue;
     }
 
-    private static String buildWebSiteContent(StringBuffer logLink) {
-        return "<html><head><title>Seatunnel log</title></head>\n"
-                + "<body>\n"
-                + " <h2>Seatunnel log</h2>\n"
-                + " <ul>\n"
-                + logLink.toString()
-                + " </ul>\n"
-                + "</body></html>";
-    }
-
-    private String getFileLogPath(PropertiesConfiguration config)
-            throws NoSuchFieldException, IllegalAccessException {
-        Field propertiesField = BuiltConfiguration.class.getDeclaredField("appendersComponent");
-        propertiesField.setAccessible(true);
-        Component propertiesComponent = (Component) propertiesField.get(config);
-        StrSubstitutor substitutor = config.getStrSubstitutor();
-        return propertiesComponent.getComponents().stream()
-                .filter(component -> "fileAppender".equals(component.getAttributes().get("name")))
-                .map(component -> substitutor.replace(component.getAttributes().get("fileName")))
-                .findFirst()
-                .orElse(null);
-    }
-
-    /** Get configuration log path */
-    private String getLogPath() throws NoSuchFieldException, IllegalAccessException {
-        String routingAppender = "routingAppender";
-        String fileAppender = "fileAppender";
-        PropertiesConfiguration config = getLogConfiguration();
-        // Get routingAppender log file path
-        String routingLogFilePath = getRoutingLogFilePath(config);
-
-        // Get fileAppender log file path
-        String fileLogPath = getFileLogPath(config);
-        String logRef =
-                config.getLoggerConfig(StringUtils.EMPTY).getAppenderRefs().stream()
-                        .map(Object::toString)
-                        .filter(ref -> ref.contains(routingAppender) || ref.contains(fileAppender))
-                        .findFirst()
-                        .orElse(StringUtils.EMPTY);
-        if (logRef.equals(routingAppender)) {
-            return routingLogFilePath.substring(0, routingLogFilePath.lastIndexOf("/"));
-        } else if (logRef.equals(fileAppender)) {
-            return fileLogPath.substring(0, routingLogFilePath.lastIndexOf("/"));
-        } else {
-            logger.warning(String.format("Log file path is empty, get logRef : %s", logRef));
-            return null;
-        }
-    }
-
     /** Get Current Node Log By /log request */
-    private void getCurrentNodeLog(HttpGetCommand httpGetCommand, String uri)
-            throws NoSuchFieldException, IllegalAccessException {
+    private void getCurrentNodeLog(HttpGetCommand httpGetCommand, String uri) {
         String logName = getParam(uri);
-        String logPath = getLogPath();
+        String logPath = logService.getLogPath();
 
         if (StringUtils.isBlank(logName)) {
             // Get Current Node Log List
-            List<File> logFileList = FileUtils.listFile(logPath);
-            StringBuffer logLink = new StringBuffer();
-            for (File file : logFileList) {
-                logLink.append(buildLogLink("log/" + file.getName(), file.getName()));
-            }
-            this.prepareResponse(httpGetCommand, getRestValue(buildWebSiteContent(logLink)));
+            this.prepareResponse(httpGetCommand, getRestValue(logService.currentNodeLog(uri)));
         } else {
             // Get Current Node Log Content
             prepareLogResponse(httpGetCommand, logPath, logName);
@@ -492,64 +330,15 @@ public class RestHttpGetCommandProcessor extends HttpCommandProcessor<HttpGetCom
         }
     }
 
-    public String buildLogLink(String href, String name) {
-        return "<li><a href=\"" + href + "\">" + name + "</a></li>\n";
-    }
+    private void getAllLogName(HttpGetCommand httpGetCommand) {
 
-    private static String sendGet(String urlString) {
         try {
-            HttpURLConnection connection = (HttpURLConnection) new URL(urlString).openConnection();
-            connection.setRequestMethod("GET");
-            connection.setConnectTimeout(5000);
-            connection.setReadTimeout(5000);
-            connection.connect();
-
-            if (connection.getResponseCode() == 200) {
-                try (InputStream is = connection.getInputStream();
-                        ByteArrayOutputStream baos = new ByteArrayOutputStream()) {
-                    byte[] buffer = new byte[1024];
-                    int len;
-                    while ((len = is.read(buffer)) != -1) {
-                        baos.write(buffer, 0, len);
-                    }
-                    return baos.toString();
-                }
-            }
-        } catch (IOException e) {
-            e.printStackTrace();
-        }
-        return null;
-    }
-
-    private void getAllLogName(HttpGetCommand httpGetCommand)
-            throws NoSuchFieldException, IllegalAccessException {
-        String logPath = getLogPath();
-        List<File> logFileList = FileUtils.listFile(logPath);
-        List<String> fileNameList =
-                logFileList.stream().map(File::getName).collect(Collectors.toList());
-        try {
-            this.prepareResponse(httpGetCommand, JsonUtils.toJsonString(fileNameList));
+            this.prepareResponse(httpGetCommand, JsonUtils.toJsonString(logService.allLogName()));
         } catch (SeaTunnelRuntimeException e) {
             httpGetCommand.send400();
-            logger.warning(String.format("Log file name get failed, get log path: %s", logPath));
+            logger.warning(
+                    String.format(
+                            "Log file name get failed, get log path: %s", logService.getLogPath()));
         }
-    }
-
-    private static String getRoutingLogFilePath(PropertiesConfiguration config)
-            throws NoSuchFieldException, IllegalAccessException {
-        Field propertiesField = BuiltConfiguration.class.getDeclaredField("appendersComponent");
-        propertiesField.setAccessible(true);
-        Component propertiesComponent = (Component) propertiesField.get(config);
-        StrSubstitutor substitutor = config.getStrSubstitutor();
-        return propertiesComponent.getComponents().stream()
-                .filter(
-                        component ->
-                                "routingAppender".equals(component.getAttributes().get("name")))
-                .flatMap(component -> component.getComponents().stream())
-                .flatMap(component -> component.getComponents().stream())
-                .flatMap(component -> component.getComponents().stream())
-                .map(component -> substitutor.replace(component.getAttributes().get("fileName")))
-                .findFirst()
-                .orElse(null);
     }
 }
