+++ seatunnel-engine/seatunnel-engine-server/src/main/java/org/apache/seatunnel/engine/server/rest/servlet/AllNodeLogServlet.java
@@ -17,18 +17,13 @@
 
 package org.apache.seatunnel.engine.server.rest.servlet;
 
-import org.apache.seatunnel.shade.com.fasterxml.jackson.databind.node.ArrayNode;
-
-import org.apache.seatunnel.common.utils.JsonUtils;
 import org.apache.seatunnel.engine.common.config.server.HttpConfig;
 import org.apache.seatunnel.engine.server.SeaTunnelServer;
 import org.apache.seatunnel.engine.server.log.FormatType;
+import org.apache.seatunnel.engine.server.rest.service.LogService;
 
 import org.apache.commons.lang3.StringUtils;
 
-import com.hazelcast.internal.json.JsonArray;
-import com.hazelcast.internal.json.JsonObject;
-import com.hazelcast.jet.datamodel.Tuple3;
 import com.hazelcast.spi.impl.NodeEngineImpl;
 import lombok.extern.slf4j.Slf4j;
 
@@ -37,16 +32,15 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
 import java.io.IOException;
-import java.util.ArrayList;
-
-import static org.apache.seatunnel.engine.server.rest.RestConstant.GET_ALL_LOG_NAME;
-import static org.apache.seatunnel.engine.server.rest.RestConstant.GET_LOGS;
 
 @Slf4j
 public class AllNodeLogServlet extends LogBaseServlet {
 
+    private final LogService logService;
+
     public AllNodeLogServlet(NodeEngineImpl nodeEngine) {
         super(nodeEngine);
+        this.logService = new LogService(nodeEngine);
     }
 
     @Override
@@ -57,71 +51,26 @@ public class AllNodeLogServlet extends LogBaseServlet {
         HttpConfig httpConfig =
                 seaTunnelServer.getSeaTunnelConfig().getEngineConfig().getHttpConfig();
         String contextPath = httpConfig.getContextPath();
-        int port = httpConfig.getPort();
         String uri = req.getRequestURI();
 
         // Analysis uri, get logName and jobId param
-        String param = getLogParam(uri, contextPath);
+        String param = logService.getLogParam(uri, contextPath);
         boolean isLogFile = param.contains(".log");
         String logName = isLogFile ? param : StringUtils.EMPTY;
         String jobId = !isLogFile ? param : StringUtils.EMPTY;
 
-        String logPath = getLogPath();
-        JsonArray systemMonitoringInformationJsonValues =
-                getSystemMonitoringInformationJsonValues();
+        String logPath = logService.getLogPath();
 
         if (StringUtils.isBlank(logName)) {
-            StringBuffer logLink = new StringBuffer();
-            ArrayList<Tuple3<String, String, String>> allLogNameList = new ArrayList<>();
-
-            systemMonitoringInformationJsonValues.forEach(
-                    systemMonitoringInformation -> {
-                        String host = systemMonitoringInformation.asObject().get("host").asString();
-                        String url = "http://" + host + ":" + port + contextPath;
-                        String allName = sendGet(url + GET_ALL_LOG_NAME);
-                        log.debug(String.format("Request: %s , Result: %s", url, allName));
-                        ArrayNode jsonNodes = JsonUtils.parseArray(allName);
-
-                        jsonNodes.forEach(
-                                jsonNode -> {
-                                    String fileName = jsonNode.asText();
-                                    if (StringUtils.isNotBlank(jobId)
-                                            && !fileName.contains(jobId)) {
-                                        return;
-                                    }
-                                    allLogNameList.add(
-                                            Tuple3.tuple3(
-                                                    host + ":" + port,
-                                                    url + GET_LOGS + "/" + fileName,
-                                                    fileName));
-                                });
-                    });
 
             FormatType formatType = FormatType.fromString(req.getParameter("format"));
             switch (formatType) {
                 case JSON:
-                    JsonArray jsonArray =
-                            allLogNameList.stream()
-                                    .map(
-                                            tuple -> {
-                                                JsonObject jsonObject = new JsonObject();
-                                                jsonObject.add("node", tuple.f0());
-                                                jsonObject.add("logLink", tuple.f1());
-                                                jsonObject.add("logName", tuple.f2());
-                                                return jsonObject;
-                                            })
-                                    .collect(JsonArray::new, JsonArray::add, JsonArray::add);
-                    writeJson(resp, jsonArray);
+                    writeJson(resp, logService.allNodeLogFormatHtml(jobId));
                     return;
                 case HTML:
                 default:
-                    allLogNameList.forEach(
-                            tuple ->
-                                    logLink.append(
-                                            buildLogLink(
-                                                    tuple.f1(), tuple.f0() + "-" + tuple.f2())));
-                    String logContent = buildWebSiteContent(logLink);
-                    writeHtml(resp, logContent);
+                    writeHtml(resp, logService.allNodeLogFormatHtml(jobId));
             }
         } else {
             prepareLogResponse(resp, logPath, logName);
